==> code/2._Data_Cleaning.ipynb <==
# Data Cleaning

This notebook aims to clean up the play-by-play dataset. There are over 200 features in the dataset, but the first step is determining which of these features are established before the play occurs vs. after the play occurs. The next step is dealing with the non-numeric variables in the data. Finally, I take a look at the nulls.

**Please Note: Restarting this notebook will cause errors in the code since all the data is not stored locally.**

## Contents
[Eliminating Post-play Features](#Eliminating-Post-play-Features)<br>
[Non-Numeric Variables](#Non-Numeric-Variables)<br>
[Dealing with Nulls](#Dealing-with-Nulls)

import pandas as pd
import s3fs
import numpy as np

pd.set_option('display.max_columns', 999)
pd.set_option('display.max_rows', 999999)

import warnings
warnings.filterwarnings('ignore')

fs = s3fs.S3FileSystem(anon=False,key='AWS KEY',secret='AWS SECRET KEY')

key = 'nfl_play_by_play_with_weather_2009_2018.csv'
bucket = 'nfl-play-by-play-capstone'

df = pd.read_csv(fs.open('{}/{}'.format(bucket, key),
                         mode='rb'))

### Eliminating Post-play Features

There are more than 250 variables in this dataset, but many of these variables are information gained after the play occurred. For example, incomplete_pass refers to whether or not a pass was completed on that play. If we are predicting play type, we can't use this information because we would not have it prior to the play. I narrowed the dataframe down to features that are considered pre-play using the definitions of variables found [here](https://github.com/maksimhorowitz/nflscrapR/blob/master/R/scrape_play_by_play.R).

df = df[['forecast', 'week', 'wind', 'year','game_id','home_team','away_team','posteam','posteam_type','defteam',
         'side_of_field','yardline_100','game_date','quarter_seconds_remaining','half_seconds_remaining',
         'game_seconds_remaining','game_half','quarter_end','drive','qtr','down','goal_to_go','time','yrdln',
         'ydstogo','desc','play_type','home_timeouts_remaining','away_timeouts_remaining',
         'posteam_timeouts_remaining','defteam_timeouts_remaining','total_home_score','total_away_score',
         'posteam_score','defteam_score','score_differential','wp']]

### Non-Numeric Variables

In this section, I want to look at the non-numeric data. I will need to do one of three things to these variables: ignore them, extract numeric data from strings, or dummy the columns. First, let's look at the non-numeric columns.

df.select_dtypes(include = 'object').columns

df.select_dtypes(include = 'object').head()

First, let's look at forecast. I want to separate the temperature into temperature and description.

df['temperature'] = df['forecast'].str.split('f ',expand=True)[0]
df['weather_desc'] = df['forecast'].str.split('f ',expand=True)[1]
df = df.drop(columns='forecast')

Now let's check the values we have for temperature.

df['temperature'].value_counts()

Already I can see that Dome will be difficult to replace or impute. I can't be sure with the information I have whether or not the Dome was open or closed during the game. There are also two games that have it seems two temperatures. I am first going update those temperatures. Then, for now I am going to create a binary column for whether or not a game was in a dome and binary columns for ranges of temperatures (cold < 40 degrees, moderate (41,70), hot > 70 degrees). To create these columns, I am going to impute 65 for temperature in Dome stadiums.

df['temperature'].replace('33/51','51',inplace=True)
df['temperature'].replace('39/53','53',inplace=True)

df['dome'] = np.where(df['temperature'] == 'DOME',1,0)

df['temperature'] = df['temperature'].replace('DOME',65).astype(float)

mask = [df['temperature'].between(-100,45), df['temperature'].between(46,70), df['ydstogo'].between(71,150)]
values = ['Cold','Moderate','Hot']
df['temperature_range'] = np.select(mask, values, 0)

df = df.join(pd.get_dummies(df['temperature_range'], prefix='temp_range',drop_first=True))
df = df.drop(columns=['temperature_range','temperature'])

Next, I need to use the weather description. I decided to create a dummy variable for whether or not there is precipitation during the game. If the forecast was a Dome, the weather description was null. I am going to make the assumption that there is no precipitation in a Dome.

df['weather_desc'].value_counts()

df['precipitation'] = df['weather_desc'].fillna('NA').apply(lambda x: 1 if ('Rain' in x 
                                                                           or 'Snow' in x 
                                                                           or 'Wintry' in x
                                                                           or 'Drizzle' in x
                                                                           or 'Fog' in x
                                                                           or 'Showers' in x
                                                                           or 'Flurries' in x
                                                                           or 'Blizzard' in x
                                                                           or 'Thunder' in x) else 0)
df = df.drop(columns='weather_desc')

Next, wind speed needs to be extracted from the wind string. NFLWeather.com did not start collecting wind consistently until about 2015. I probably won't use this feature since so many values are missing (30%), but I will still extract the number for what it is worth.

df[df['wind'].isnull()]['year'].max()

df['wind'].isnull().mean()

df['wind_speed'] = df['wind'].str.split('m ', expand=True)[0].astype(float)
df = df.drop(columns='wind')

Pos team & def team will be dummied, but the dummying will created more than 60 additional columns. For storage purposes, I will dummy these in the feature engineering notebook. The next variable to address is whether or not the possession team is home.

df['posteam_home'] = df['posteam_type'].apply(lambda x: 1 if x == 'home' else 0)
df = df.drop(columns=['home_team','away_team','posteam_type'])

I am also going to make the game_date a datetime object rather thana a string.

df['game_date'] = pd.to_datetime(df['game_date'])

Next, I will binarize whether or not the team is on their own side.

df['own_side'] = np.where(df['posteam'] == df['side_of_field'],1,0)
df = df.drop(columns=['side_of_field'])

Next, I am going to eliminate plays in overtime and binarize whether or not it is the second half.

df['game_half'].value_counts()

df = df[df['game_half'] != 'Overtime']

df['second_half'] = df['game_half'].apply(lambda x: 1 if x =='Half2' else 0)
df = df.drop(columns='game_half')

Next I am going to drop time since we already have quarter_seconds_left, half_seconds_left, and game_seconds_left.

df = df.drop(columns = ['time'])

Finally, I am going to drop the column yrdln because we already have yardline_100 with essentially the same information.

df = df.drop(columns='yrdln')

df.select_dtypes(include = 'object').head()

As I mentioned above, posteam and defteam will be dummied in the feature engineering notebook. Play description will remain, but I will not use it in the model. Playtype will be what we are predicting, so no need to change at the moment.

### Dealing with Nulls

In this section, I aim to address any nulls we have in the data. First I need to isolate the plays to those that are effectively a run (run or qb kneel) or effectively a pass (pass or qb spike).

df.head()

df['play_type'].value_counts()

df = df[(df['play_type'] == 'pass') | (df['play_type'] == 'run') | (df['play_type'] == 'qb_kneel')
       | (df['play_type'] == 'qb_spike')]

df['play_type'].value_counts()

df['effective_run'] = df['play_type'].map({'pass': 0, 'qb_spike':0, 'run':1, 'qb_kneel':1})

df['effective_pass'] = df['play_type'].map({'pass': 1, 'qb_spike':1, 'run':0, 'qb_kneel':0})

The next step is addressing the nulls.

df.isnull().sum()

First, we need to impute half_seconds_remaining & game_seconds_remaining. We have enough information given quarter_seconds_remaining and qtr. In order to impute these, I need to reindex the dataframe since we just dropped a bunch of play types.

df.reset_index(drop=True,inplace=True)

Next, I will impute half_seconds_remaining.

for i in df[df['half_seconds_remaining'].isnull()].index:
    seconds = df.iloc[i]['quarter_seconds_remaining'] + 900
    df.loc[[i],['half_seconds_remaining']] = seconds

Next, I will impute game_seconds_remaining.

for i in df[df['game_seconds_remaining'].isnull()].index:
    if (df.iloc[i]['qtr'] == 3):
        seconds = df.iloc[i]['half_seconds_remaining']
        df.loc[[i],['game_seconds_remaining']] = seconds
    elif (df.iloc[i]['qtr'] == 1) | (df.iloc[i]['qtr'] == 2):
        seconds = df.iloc[i]['half_seconds_remaining'] + 1800
        df.loc[[i],['game_seconds_remaining']] = seconds
    else:
        pass

Next, I will take a look at the missing downs.

df[df['down'].isnull()].head()

It appears that each play with a missing down is a two-point conversion, a botched PAT attempt, or some kickoffs were labeld as qb kneels in the endzone. Let's check.

counter = 0
for row in df['desc']:
    if 'TWO-POINT' in row:
        counter +=1
counter == df[df['down'].isnull()].shape[0] - df[(df['down'].isnull()) & ~(df['desc'].str.contains('TWO-POINT')) &
                                                 df['desc'].str.contains('kicks')].shape[0]

df[(df['down'].isnull()) & ~(df['desc'].str.contains('TWO-POINT'))].shape[0]/df.shape[0]

All downs that are missing are kickoffs or two-point conversions. For the purposes of this project, I am going to drop the kickoffs & two-point conversions (less than 0.02% of plays).

df = df.dropna(subset=['down'])

Next, we need to look posteam_score, defteam_score, and score_differential. The drops above seem to have gotten rid of those.

df[df['posteam_score'].isnull()].shape

df[df['defteam_score'].isnull()].shape

df[df['score_differential'].isnull()].shape

Next, I am going to drop where win probability is null since that calculation was done outside of this project.

df = df.dropna(subset=['wp'])

The only nulls we have left are wind_speed which simply was not collected by the source NFLWeather.com consistently until 2015. For now, we are going to leave the nulls and ignore the variable.

df.isnull().sum().sort_values(ascending=False).head()

Finally, let's save the clean csv back to S3.

fs = s3fs.S3FileSystem(anon=False,key='AWS KEY',secret='AWS SECRET KEY')

bucket = 'nfl-play-by-play-capstone'

with fs.open(f'{bucket}/nfl_play_by_play_with_weather_2009_2018_clean.csv','w') as f:
    df.to_csv(f)
