==> code/6._Modeling_for_App.ipynb <==
# Modeling for App

In this section, I aim to simplify the model and "pickle" said model for a simple application demo of the model. We want to simplify the model because the application will require a user to input features. If the user has to enter a large number of features, the application is more frustrating and less useful.

**Please Note: Restarting this notebook will cause errors in the code since all the data is not stored locally.**

### Contents

[Simplifying the Model](#Simplifying-the-Model)<br>
[Pickling Model](#Pickling-Model)

import pandas as pd
import s3fs

from sklearn.model_selection import train_test_split
import xgboost as xgb

import pickle

import warnings
warnings.filterwarnings("ignore")

fs = s3fs.S3FileSystem(anon=False,key='AWS KEY',secret='AWS SECRET KEY')

key = 'nfl_play_by_play_with_weather_model.csv'
bucket = 'nfl-play-by-play-capstone'

df = pd.read_csv(fs.open('{}/{}'.format(bucket, key),
                         mode='rb')).drop(columns=['Unnamed: 0'])

### Simplifying the Model

In this section, I will choose a select few features that have the most predictive power. The original model has a little over 100 features (about 60 of which just indicate who is one offense and who is on defense). Reducing the features to only 8 achieves almost exactly the same accuracy.

X = df[['down','qtr','game_seconds_remaining','half_seconds_remaining','quarter_seconds_remaining','ydstogo',
        'score_differential','yardline_100']].values
y = df['effective_run']
X_train, X_test, y_train, y_test = train_test_split(X,y, random_state=42)

xgb_model = xgb.XGBClassifier(objective="binary:logistic", random_state=42, 
                              eval_metric=["error"], max_depth = 8, subsample=0.7)
xgb_model.fit(X_train, y_train, early_stopping_rounds=5,eval_set=[(X_test, y_test)])

print(f"Test Accuracy Score: {xgb_model.score(X_test,y_test)}")
print(f"Train Accuracy Score: {xgb_model.score(X_train,y_train)}")

As you can see, the simple model achieves almost the same accuracy with much less information. The less inputs we have, the easier the application will run.

### Pickling Model

Pickling the model is only one line of code, but essentially this stores the model so we can call it elsewhere and make predictions.

pickle.dump(xgb_model, open('../app/model.p', 'wb'))
